+++ aa_customizer/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "1.1.2a"
+__version__ = "1.1.3"
 __title__ = "Customizer"